requirements.txt
Flask==3.1.2
firebase-admin==7.1.0
python-dotenv==1.2.1
requests==2.32.5
pytest==9.0.1
flask-cors==6.0.1
gunicorn>=20.0
